float fix breakdown

+++ evaluator/views.py
@@ -1088,8 +1088,8 @@ class MakeQuatationPhase1(IsEvaluator,View):
 		
 		payment_method = request.POST.get('payment_method')
 		attender_notes = request.POST.get('attender_notes')
-		before_cleaning_amount	= int(request.POST.get('before_cleaning_amount')or 0)
-		after_cleaning_amount	= int(request.POST.get('after_cleaning_amount')or 0)
+		before_cleaning_amount	= float(request.POST.get('before_cleaning_amount')or 0.000)
+		after_cleaning_amount	= float(request.POST.get('after_cleaning_amount')or 0.000)
 
 		#update payment method
 		Evaluation.objects.filter(id=evaluation_id,is_active=True).update(payment_method=payment_method,attender_notes=attender_notes,quatation_status='PENDING',before_cleaning_amount=before_cleaning_amount,after_cleaning_amount=after_cleaning_amount)
